# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts for the org
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactWrite"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a single contact
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactWrite"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /market-locations:
    get:
      operationId: listMarketLocations
      tags: [marketLocations]
      summary: List market locations for the org
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      responses:
        "200":
          description: Market location list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketLocation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createMarketLocation
      tags: [marketLocations]
      summary: Create a market location
      parameters:
        - $ref: "#/components/parameters/XOrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketLocationWrite"
      responses:
        "201":
          description: Created market location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketLocation"
  /market-locations/{id}:
    patch:
      operationId: updateMarketLocation
      tags: [marketLocations]
      summary: Update a market location's metadata
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - name: id
          in: path
          required: true
          description: locationKey (e.g. loc-berlin)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketLocationPatch"
      responses:
        "200":
          description: Updated market location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketLocation"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteMarketLocation
      tags: [marketLocations]
      summary: Delete a market location
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /market-locations/{id}/benchmarks:
    post:
      operationId: refreshMarketBenchmarks
      tags: [marketLocations]
      summary: Upsert benchmarks (replace by usageType) and append an update log entry
      parameters:
        - $ref: "#/components/parameters/XOrgId"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketBenchmarksRefresh"
      responses:
        "200":
          description: Updated market location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketLocation"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    XOrgId:
      name: X-Org-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient role or wrong org
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        category:
          type: string
        subcategory:
          type: string
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        linkedObjectIds:
          type: array
          items:
            type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - category
        - firstName
        - lastName
        - createdAt
        - updatedAt
    ContactWrite:
      type: object
      properties:
        category:
          type: string
        subcategory:
          type: string
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        linkedObjectIds:
          type: array
          items:
            type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
      required:
        - category
        - firstName
        - lastName
...
    MarketBenchmark:
      type: object
      properties:
        id:
          type: string
        locationId:
          type: string
        usageType:
          type: string
        rentMin:
          type: number
        rentMax:
          type: number
        rentMedian:
          type: number
        purchasePriceMin:
          type: number
        purchasePriceMax:
          type: number
        purchasePriceMedian:
          type: number
        multiplierMin:
          type: number
        multiplierMax:
          type: number
        multiplierMedian:
          type: number
        vacancyRatePercent:
          type: number
        confidenceScore:
          type: number
        sourceLabel:
          type: string
        lastUpdated:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - id
        - locationId
        - usageType
        - rentMin
        - rentMax
        - rentMedian
        - purchasePriceMin
        - purchasePriceMax
        - purchasePriceMedian
        - multiplierMin
        - multiplierMax
        - multiplierMedian
        - vacancyRatePercent
        - confidenceScore
        - sourceLabel
        - lastUpdated
    MarketUpdateEntry:
      type: object
      properties:
        id:
          type: string
        locationId:
          type: string
        timestamp:
          type: string
        updatedBy:
          type: string
        changes:
          type: string
        sourceLabel:
          type: string
      required:
        - id
        - locationId
        - timestamp
        - updatedBy
        - changes
        - sourceLabel
    MarketLocation:
      type: object
      properties:
        id:
          type: string
          description: Stable client-side location key (e.g. loc-berlin).
        city:
          type: string
        region:
          type: string
        submarket:
          type: string
        lastUpdated:
          type: string
        benchmarks:
          type: array
          items:
            $ref: "#/components/schemas/MarketBenchmark"
        updateLog:
          type: array
          items:
            $ref: "#/components/schemas/MarketUpdateEntry"
      required:
        - id
        - city
        - region
        - submarket
        - lastUpdated
        - benchmarks
        - updateLog
    MarketLocationWrite:
      type: object
      properties:
        id:
          type: string
        city:
          type: string
        region:
          type: string
        submarket:
          type: string
        lastUpdated:
          type: string
      required:
        - id
        - city
        - region
        - submarket
    MarketLocationPatch:
      type: object
      properties:
        city:
          type: string
        region:
          type: string
        submarket:
          type: string
        lastUpdated:
          type: string
    MarketBenchmarksRefresh:
      type: object
      properties:
        benchmarks:
          type: array
          items:
            $ref: "#/components/schemas/MarketBenchmark"
        updateEntry:
          $ref: "#/components/schemas/MarketUpdateEntry"
      required:
        - benchmarks
        - updateEntry
...